sdk/python/requirements.txt
requests==2.22.0
PyJWT==1.7.1